Update hardware.sh - BUGFIX - Removes extra \newlines when querying media drives (CF card, SD Card)
This update fixes a long standing bug that existed when a computer had a media drive (such as an SD card reader) built-in. The drive would get detected as a /dev/sd* device, and extra \newline lines would be added to the LaTeX file, pushing the DVD drive section, and other sections further down the page. In some cases this would result in specs being on more than 1 page. This fixes the issue.